Move disableHttpLogging to BuildLoggers
