+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/BuildImageTask.java
@@ -105,7 +105,7 @@ public class BuildImageTask extends DefaultTask {
             .build();
 
     // TODO: Instead of disabling logging, have authentication credentials be provided
-    GradleProjectProperties.disableHttpLogging();
+    GradleBuildLogger.disableHttpLogging();
 
     RegistryClient.setUserAgentSuffix(USER_AGENT_SUFFIX);
