+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/BuildImageMojo.java
@@ -105,7 +105,7 @@ public class BuildImageMojo extends JibPluginConfiguration {
             .build();
 
     // TODO: Instead of disabling logging, have authentication credentials be provided
-    MavenProjectProperties.disableHttpLogging();
+    MavenBuildLogger.disableHttpLogging();
 
     RegistryClient.setUserAgentSuffix(USER_AGENT_SUFFIX);
