+++ jib-maven-plugin/src/main/java/com/google/cloud/tools/jib/maven/MavenBuildLogger.java
@@ -22,6 +22,13 @@ import org.apache.maven.plugin.logging.Log;
 /** Implementation of {@link BuildLogger} for Maven plugins. */
 class MavenBuildLogger implements BuildLogger {
 
+  /** Disables annoying Apache HTTP client logging. */
+  static void disableHttpLogging() {
+    System.setProperty(
+        "org.apache.commons.logging.Log", "org.apache.commons.logging.impl.SimpleLog");
+    System.setProperty("org.apache.commons.logging.simplelog.defaultlog", "error");
+  }
+
   private final Log log;
 
   MavenBuildLogger(Log log) {
